a2a_agents/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
a2a-sdk==0.3.8
httpx>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
